Rename CoreConfig to NodeConfig and expose it
